requirements.txt
flask
pypdf
python-docx
pandas
openpyxl
xlrd
beautifulsoup4
python-pptx
tqdm
pytest
